Sonar: proper numeric assignment
Turn some of my indices to the same index type as in domain (sup::dto::uint32 for job, instruction indices)

+++ src/lib/oac-tree-gui-components/oac_tree_gui/jobsystem/abstract_domain_runner.cpp
@@ -85,13 +85,13 @@ void AbstractDomainRunner::Reset()
   m_domain_job->Reset();
 }
 
-void AbstractDomainRunner::SetBreakpoint(std::size_t instr_idx)
+void AbstractDomainRunner::SetBreakpoint(std::uint32_t instr_idx)
 {
   ValidateJob();
   m_domain_job->SetBreakpoint(instr_idx);
 }
 
-void AbstractDomainRunner::RemoveBreakpoint(std::size_t instr_idx)
+void AbstractDomainRunner::RemoveBreakpoint(std::uint32_t instr_idx)
 {
   ValidateJob();
   m_domain_job->RemoveBreakpoint(instr_idx);
