+++ src/lib/oac-tree-gui-components/oac_tree_gui/jobsystem/automation_client.h
@@ -57,7 +57,7 @@ public:
 
   std::size_t GetJobCount() const override;
 
-  std::string GetProcedureName(std::size_t job_index) const override;
+  std::string GetProcedureName(std::uint32_t job_index) const override;
 
   std::unique_ptr<AbstractJobHandler> CreateJobHandler(RemoteJobItem* job_item,
                                                        const UserContext& user_context) override;
