+++ src/lib/oac-tree-gui-components/oac_tree_gui/jobsystem/remote_domain_runner.cpp
@@ -32,7 +32,7 @@ namespace oac_tree_gui
 RemoteDomainRunner::RemoteDomainRunner(DomainEventDispatcherContext dispatcher_context,
                                        UserContext user_context,
                                        sup::oac_tree_server::IJobManager& manager,
-                                       std::size_t job_index)
+                                       std::uint32_t job_index)
     : AbstractDomainRunner(std::move(dispatcher_context), std::move(user_context))
 {
   auto remote_job = sup::oac_tree_server::CreateClientJob(
